Efficiency bonus of basic, standard, advancec
As CP costs of standard and advanced are the same as basic when they are also combo'd, there is no point in casting basic or standard when in the middle of a combo.
Profiler time increased but for longer search decreased. For the first step Preparatory touch, long search time was about 8:00, with combo checking about 4:50