Repair "Common Sense" logic
Made it so that there is a minimum threshold for a repair. This number could be changed but it's more of a general rule to not repair unless you can get the maximum out of it: which is usually below 15 durability.
Profiler time to first step 23s. All steps took 14:28